refactor: centralize types
