+++ murmer_client/src/lib/stores/voiceUsers.ts
@@ -1,6 +1,6 @@
 import { writable } from 'svelte/store';
 import { chat } from './chat';
-import type { Message } from './chat';
+import type { Message } from '../types';
 
 function createVoiceUserStore() {
   const { subscribe, set } = writable<string[]>([]);
