+++ murmer_client/src/lib/types.ts
@@ -1,0 +1,19 @@
+export interface Message {
+  type: string;
+  user: string;
+  text?: string;
+  time?: string;
+  [key: string]: unknown;
+}
+
+export interface RemotePeer {
+  id: string;
+  stream: MediaStream;
+  stats?: ConnectionStats;
+}
+
+export interface ConnectionStats {
+  rtt: number;
+  jitter: number;
+  strength: number;
+}
